Add files via upload
Well, here's an interesting project you may like. BookNFlix a website where you can search and watch movies and books that are recommended by collaborative filtering. and in the future the search bar will be working 🤞. and here's the link to the working website on fire base= https://booksnflix-version-1.netlify.app/